chore(get): rename `https` to `request`
